+++ src/fullInput.txt
@@ -1,4 +1,4 @@
-5000
+1
 54408008 Randall Ashley 118 7386823466
 78724186 Janiya Wang 261 3316810536
 78786820 Joselyn Peters 070 0653736007
