+++ src/main.cpp
@@ -1,5 +1,5 @@
-#pragma once
-#include "laundryBackend.cpp"
+
+#include "C:\\Users\\suyda\\GitHub\\Laundry-Board-Project\\src\\laundryBackend.cpp"
 #include <iostream>
 #include <regex>
 #include <string>
@@ -23,7 +23,9 @@ int main()
 
     int n = 0;
     inputFile >> n;
-    AVLTree userTree;
+    //AVLTree laundryTree;
+    MapImplementation laundryMap;
+    //SetImplementation laundrySet;
     for (int i = 0; i < n; i++)
     {
         // read in the next line and parse it into a constructor for a user node
@@ -32,43 +34,58 @@ int main()
         string fullName = "";
         string phone = "";
         string gatorID = "";
-        string roomNum = "";
+        string room = "";
 
         inputFile >> gatorID;
         inputFile >> firstName;
         inputFile >> lastName;
-        inputFile >> roomNum;
+        inputFile >> room;
         inputFile >> phone;
 
         int gator1ID = stoi(gatorID);
-        int roomNumber = stoi(roomNum);
+        int roomNum = stoi(room);
         long phoneNum = 0;
         fullName = firstName + " " + lastName;
         
         // push that to the tree
-        userTree.setRoot(userTree.insertFromFile(userTree.getRoot(), gator1ID, fullName, phoneNum, roomNumber));
+        // laundryTree.setRoot(laundryTree.insertFromFile(laundryTree.getRoot(), gator1ID, fullName, phoneNum, roomNum));
+
+        // push that to the map
+        MapImplementation::MapUser newMapUser(gator1ID, fullName, phoneNum, roomNum);
+        laundryMap.map.emplace(i, newMapUser);
+
+        // push that to the set
+       // SetImplementation::SetUser newSetUser(gator1ID, fullName, phoneNum, roomNum);
+       // laundrySet.mset.emplace(i, newSetUser);
+
         
     }
 
     //Main Variables
-    string userInput = "";
+    string userInput;
     bool running = true;
-    vector<double> info(5, 0.0);
+    vector<double> AVLInfo(5, 0.0);
+    vector<double> mapInfo(5, 0.0);
+    vector<double> setInfo(5, 0.0);
     while (running)
     {
-        cin >> userInput;
+        getline(cin,userInput);
         vector<int> commands = comprehend(userInput);
         int command = commands.at(0);
         switch (command)
         {
         case 1:
         {
-            info.at(0) = totalAverageWashing(userTree.getRoot());
+            //AVLInfo.at(0) = totalAverageWashing(laundryTree.getRoot());
+            mapInfo.at(0) = laundryMap.totalAverageWashing();
+           // setInfo.at(0) = laundrySet.totalAverageWashing();
             break;
         }
         case 2:
         {
-            info.at(1) = totalAverageDrying(userTree.getRoot());
+            //AVLInfo.at(1) = totalAverageDrying(laundryTree.getRoot());
+            mapInfo.at(1) = laundryMap.totalAverageDrying();
+           // setInfo.at(1) = laundrySet.totalAverageDrying();
             break;
         }
         case 3:
@@ -79,25 +96,47 @@ int main()
             int userToFind = commands.at(1);
 
             //Get all of the nodes in the tree
-            vector<AVLTree::UserNode*> nodes = userTree.inorderTraversal(userTree.getRoot(), 0);
-            AVLTree::UserNode* userToCalc = nullptr;
+            //vector<AVLTree::UserNode*> nodes = laundryTree.inorderTraversal(laundryTree.getRoot(), 0);
+            //AVLTree::UserNode* avlUserToCalc = nullptr;
 
-            //find the node that you want to calculate if it exists;
-            for (int i = 0; i < nodes.size(); i++)
+            //get all of the nodes in the map
+            MapImplementation::MapUser* mapUserToCalc = &laundryMap.map.at(commands.at(1));
+            mapUserToCalc->userAverageWashing(mapUserToCalc);
+            //Get all of the nodes in the set
+            //vector<SetImplementation::SetUser*> setUsers = laundrySet.generateSetVector();
+            //SetImplementation::SetUser* setUserToCalc = nullptr;
+            /*for (int i = 0; i < setUsers.size(); i++)
             {
-                if(nodes.at(i)->getGator1ID() == userToFind)
+                if(setUsers.at(i)->getGator1ID() == userToFind)
                 {
-                    userToCalc = nodes.at(i);
+                    setUserToCalc = setUsers.at(i);
                 }
             }
-            if(userToCalc != nullptr)
+            if(setUserToCalc != nullptr)
             {
-                userToCalc->setWashAverage(userToCalc->userAverageWashing(userToCalc));
+                setUserToCalc->setWashAverage(setUserToCalc->userAverageWashing());
             }
             else
             {
                 cout << "ERRRIRRRAGJKRGkenrgvabjlaejlaj" <<endl;
-            }
+            }*/
+
+            //find the node that you want to calculate if it exists;
+            // for (int i = 0; i < nodes.size(); i++)
+            // {
+            //     if(nodes.at(i)->getGator1ID() == userToFind)
+            //     {
+            //         avlUserToCalc = nodes.at(i);
+            //     }
+            // }
+            // if(avlUserToCalc != nullptr)
+            // {
+            //     avlUserToCalc->setWashAverage(avlUserToCalc->userAverageWashing(avlUserToCalc));
+            // }
+            // else
+            // {
+            //     cout << "ERRRIRRRAGJKRGkenrgvabjlaejlaj" <<endl;
+            // }
             
             break;
         }
@@ -110,34 +149,57 @@ int main()
             int userToFind = commands.at(1);
 
             //Get all of the nodes in the tree
-            vector<AVLTree::UserNode*> nodes = userTree.inorderTraversal(userTree.getRoot(), 0);
-            AVLTree::UserNode* userToCalc = nullptr;
+            //vector<AVLTree::UserNode*> nodes = laundryTree.inorderTraversal(laundryTree.getRoot(), 0);
+            //AVLTree::UserNode* avlUserToCalc = nullptr;
 
-            //find the node that you want to calculate if it exists;
-            for (int i = 0; i < nodes.size(); i++)
+            //get all of the nodes in the map
+            MapImplementation::MapUser* mapUserToCalc = &laundryMap.map.at(commands.at(1));
+            mapUserToCalc->userAverageWashing(mapUserToCalc);
+            //Get all of the nodes in the set
+            //vector<SetImplementation::SetUser*> setUsers = laundrySet.generateSetVector();
+            /*SetImplementation::SetUser* setUserToCalc = nullptr;
+            for (int i = 0; i < setUsers.size(); i++)
             {
-                if(nodes.at(i)->getGator1ID() == userToFind)
+                if(setUsers.at(i)->getGator1ID() == userToFind)
                 {
-                    userToCalc = nodes.at(i);
+                    setUserToCalc = setUsers.at(i);
                 }
             }
-            if(userToCalc != nullptr)
+            if(setUserToCalc != nullptr)
             {
-                userToCalc->setDryAverage(userToCalc->userAverageDrying(userToCalc));
+                setUserToCalc->setWashAverage(setUserToCalc->userAverageWashing());
             }
             else
             {
                 cout << "ERRRIRRRAGJKRGkenrgvabjlaejlaj" <<endl;
-            }
+            }*/
+
+
+            //find the node that you want to calculate if it exists;
+            // for (int i = 0; i < nodes.size(); i++)
+            // {
+            //     if(nodes.at(i)->getGator1ID() == userToFind)
+            //     {
+            //         userToCalc = nodes.at(i);
+            //     }
+            // }
+            // if(userToCalc != nullptr)
+            // {
+            //     userToCalc->setDryAverage(userToCalc->userAverageDrying(userToCalc));
+            // }
+            // else
+            // {
+            //     cout << "ERRRIRRRAGJKRGkenrgvabjlaejlaj" <<endl;
+            // }
             break;
         }
         case 5:
         {
-            vector<int> stuff = userTree.inorderTraversal(userTree.getRoot());
-            vector<int> washingTimes = inorderTraversalWashingTimes(userTree.getRoot());
-            vector<int> dryingTimes = inorderTraversalDrying(userTree.getRoot());
+            // vector<int> stuff = laundryTree.inorderTraversal(laundryTree.getRoot());
+            // vector<int> washingTimes = inorderTraversalWashingTimes(laundryTree.getRoot());
+            // vector<int> dryingTimes = inorderTraversalDrying(laundryTree.getRoot());
 
-            thresHolds(washingTimes, dryingTimes);
+            //thresHolds(washingTimes, dryingTimes);
             break;
         }
 
@@ -154,28 +216,28 @@ int main()
 
 vector<int> comprehend(string userInput)
 {
-    if(userInput == "Wash average")
+    if(userInput.compare("Wash average")==0)
     {
         return {1};
     }
-    if(userInput == "Dry average")
+    if(userInput.compare("Dry average")==0)
     {
         return {2};
     }
-    if(userInput == "Wash average - user")
+    if(userInput.compare("Wash average - user")==0)
     {
         int userToFind = 0;
         cin >> userToFind;
         return {3, userToFind};
         
     }
-    if(userInput == "Dry average - user")
+    if(userInput.compare("Dry average - user")==0)
     {
         int userToFind = 0;
         cin >> userToFind;
         return {4, userToFind};
     }
-    if(userInput == "Thresholds")
+    if(userInput.compare("Thresholds")==0)
     {
         return {5};
     }
@@ -242,6 +304,6 @@ void thresHolds(vector<int> washing, vector<int> drying)
     
 
 
-    ///will code in after deciding on final threshold/data stuff
+
 }
